backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
requests==2.31.0
yfinance==0.2.43
pandas==2.2.2
numpy==1.26.4
openai==1.40.0